Image folder structure, optimisation, image paths
# Image structure

Images are stored in the directory `images` in the corresponding post directory. The post directory is the same as the id of the post.

That way we can source an image by using the {{ page.id }} Lyquid attribute.

# Optimization

A Rakefile and rake tasks have been added.

The `resize` task uses rmagick-ImageMagick to resize the images to a maximum of 1280x960.

The `optimizeimages` task aims to remove some excess data form files but it is slow and CPU intensive without much effect (only 6.3% improvement).

Resized/Optimised images are stored in an `optimized` directory inside each post's image directory. That way wee can source the optimized image with `![](/images{{ page.id }}/optimized/image.jpg)